[input] Refactor vkbd handling.
Trying to get some abstraction so it is easier to understand what does
what... The goal here is to handle VKBD attachement through DBus RPC.
There are a lot of state variable checked or not at different stage of
the input server.
Introduce a few domain_ handler to interface handling, then handle the
Vkbd through those.
This should manage Vkbd without modifying the rest.

OXT-544

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>